Add config file, attachtab, and locker.py support
A large rewrite to add support for a configuration file, an "attachtab"
to present mount information to userspace tools, and using locker.py for
locker lookup and handling.

Changes as part of this commit:
- IOErrors are no longer raised during lookup, instead None is returned
  and errors are syslogged.
- Whether or not to syslog for ERR lockers, unmountable lockers, or
  successful mounts is configurable in the file.
- We override Fuse.parse() and store the mountpoint for later access
- Mount data is now stored in an attachtab object which stores both
  user-requested symlinks and Locker objects
- Use new-style format strings in creating the readme

N.B. The attachtab does not expose other users' mounts, as the old
liblocker did.  The decision to not expose other users' mounts in
pyhesiodfs was deliberate, and not simply a side-effect of
implementation.